refactor(C4): extract AgentRequestBus + ChatInterventionBus → runtime/session_buses.py (#1792)
FP-0044 C-series stage 4. Extracts the two Session-backed intervention bus
adapters verbatim from `reyn.runtime.session` (the god-file) into a new
`reyn.runtime.session_buses` module (name covers both buses).

- Pure move, byte-identical per-class bodies (AgentRequestBus +
  ChatInterventionBus each verified removed==added via diff).
- `session_buses.py` docstring is current-state only (what the module IS);
  this refactor narrative lives here in the commit (C-series discipline, C1
  #1790 R1).
- `Session` (constructor arg) + `UserIntervention` / `InterventionAnswer`
  (method annotations) imported under TYPE_CHECKING only — never at runtime,
  so the session <-> session_buses cycle is broken by construction; bodies are
  duck-typed via `self._session`. One-directional `session -> session_buses`
  at runtime (session imports both buses back for its own instantiation:
  AgentRequestBus(self) + the per-spawn ChatInterventionBus factories).
- Clean break, no re-export shim: all 6 consumer files repointed
  `from reyn.runtime.session import <Bus>`
  → `from reyn.runtime.session_buses import <Bus>`.
- Repo-wide straggler 0 across all ref classes INCLUDING multi-line import
  blocks: two consumers imported `ChatInterventionBus` via the
  `from reyn.runtime.session import (\n  ...,\n  ChatInterventionBus,\n)`
  parenthesised form, which a single-line grep/repoint misses. Scan now parses
  parenthesised blocks too; verified 0 session-located bus refs remain.
