+++ tests/test_a2a_restart_resume_292.py
@@ -55,9 +55,9 @@ from reyn.core.events.state_log import StateLog
 from reyn.interfaces.web.run_registry import RunRegistry
 from reyn.runtime.session import (
     DEFAULT_CHAT_CHANNEL_ID,
-    ChatInterventionBus,
     Session,
 )
+from reyn.runtime.session_buses import ChatInterventionBus
 from reyn.user_intervention import (
     InterventionAnswer,
     InterventionChoice,
