+++ tests/test_intervention_agent_routing.py
@@ -28,7 +28,8 @@ from pathlib import Path
 
 import pytest
 
-from reyn.runtime.session import AgentRequestBus, Session
+from reyn.runtime.session import Session
+from reyn.runtime.session_buses import AgentRequestBus
 from reyn.user_intervention import InterventionAnswer, UserIntervention
 
 # ── 1. Hook methods exist with the canonical signatures ────────────────
