+++ tests/test_intervention_bus_protocols.py
@@ -36,7 +36,8 @@ import inspect
 from pathlib import Path
 
 from reyn.interfaces.web.a2a_intervention import A2AInterventionBus
-from reyn.runtime.session import ChatInterventionBus, Session
+from reyn.runtime.session import Session
+from reyn.runtime.session_buses import ChatInterventionBus
 from reyn.user_intervention import (
     InterventionAnswer,
     InterventionBus,
